+++ bin/faucet/src/config.rs
@@ -34,13 +34,15 @@ pub struct FaucetConfig {
     /// Optional: Endpoint of the remote transaction prover in the format
     /// `<protocol>://<host>[:<port>]`
     pub remote_tx_prover_url: Option<Url>,
+    /// The salt to be used by the server to generate the `PoW` seed
+    pub pow_salt: String,
 }
 
 impl Display for FaucetConfig {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!(
-            "{{ endpoint: \"{}\", node_url: \"{}\", timeout_ms: \"{}\", asset_amount_options: {:?}, faucet_account_path: \"{}\", remote_tx_prover_url: \"{:?}\" }}",
-            self.endpoint, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display(), self.remote_tx_prover_url
+            "{{ endpoint: \"{}\", node_url: \"{}\", timeout_ms: \"{}\", asset_amount_options: {:?}, faucet_account_path: \"{}\", remote_tx_prover_url: \"{:?}\", pow_salt: \"{}\" }}",
+            self.endpoint, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display(), self.remote_tx_prover_url, self.pow_salt
         ))
     }
 }
@@ -57,6 +59,7 @@ impl Default for FaucetConfig {
             asset_amount_options: AssetOptions::new(vec![100, 500, 1_000]).unwrap(),
             faucet_account_path: DEFAULT_FAUCET_ACCOUNT_PATH.into(),
             remote_tx_prover_url: None,
+            pow_salt: rand::random::<[u8; 32]>().into_iter().map(|b| b as char).collect(),
         }
     }
 }