+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Faucet now supports the usage of a remote transaction prover (#830).
 - [BUGFIX] Prevents duplicated note IDs (#842).
+- Added a required Proof-of-Work in the faucet to request tokens (#831).
 
 ## v0.8.2 (2025-05-04)